requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pymongo==4.6.0
motor==3.3.2
crewai==0.11.0
langchain-openai==0.0.2
python-dotenv==1.0.0
pydantic==2.7.4
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
feedparser==6.0.10
yfinance==0.2.28
newsapi-python==0.2.6
setuptools>=80.0.0
email-validator==2.2.0 